requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.6
python-dotenv==1.0.1
langchain-ollama==0.2.2
pdf2image==1.17.0
python-dotenv==1.0.1